requirements.txt
eth-ape>=0.2.7
pytest>=6.2.5,<7.0.0
black>=21.5
project-nft==0.1.0a1
pynata==0.1.0a1
requests==2.27.1
Flask==2.0.2
htmlmin==0.1.12
